pyproject.toml
[project]
name = "tool-use-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "appnope==0.1.4",
    "asttokens==3.0.1",
    "attrs==25.4.0",
    "brotli==1.2.0",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "comm==0.2.3",
    "dataclasses-json==0.6.7",
    "ddgs==9.10.0",
    "debugpy==1.8.19",
    "decorator==5.2.1",
    "distro==1.9.0",
    "duckduckgo-search==8.1.1",
    "executing==2.2.1",
    "fake-useragent==2.2.0",
    "frozenlist==1.8.0",
    "googleapis-common-protos==1.72.0",
    "grpcio==1.76.0",
    "h11==0.16.0",
    "h2==4.3.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "hyperframe==6.1.0",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "ipykernel==7.1.0",
    "ipython==9.9.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jiter==0.12.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jupyter-client==8.8.0",
    "jupyter-core==5.9.1",
    "langchain==1.2.3",
    "langchain-classic==1.0.1",
    "langchain-community==0.4.1",
    "langchain-core==1.2.7",
    "langchain-google-genai>=4.1.2",
    "langchain-openai==1.1.7",
    "langchain-text-splitters==1.1.0",
    "langgraph==1.0.6",
    "langgraph-checkpoint==4.0.0",
    "langgraph-prebuilt==1.0.6",
    "langgraph-sdk==0.3.3",
    "langsmith==0.6.2",
    "lxml==6.0.2",
    "marshmallow==3.26.2",
    "matplotlib-inline==0.2.1",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "numpy==2.4.1",
    "openai==2.15.0",
    "opentelemetry-api==1.39.1",
    "opentelemetry-exporter-otlp-proto-common==1.39.1",
    "opentelemetry-exporter-otlp-proto-http==1.39.1",
    "opentelemetry-instrumentation==0.60b1",
    "opentelemetry-proto==1.39.1",
    "opentelemetry-sdk==1.39.1",
    "opentelemetry-semantic-conventions==0.60b1",
    "orjson==3.11.5",
    "ormsgpack==1.12.1",
    "packaging==25.0",
    "parso==0.8.5",
    "pexpect==4.9.0",
    "platformdirs==4.5.1",
    "primp==0.15.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "protobuf==6.33.4",
    "psutil==7.2.1",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "socksio==1.0.0",
    "sqlalchemy==2.0.45",
    "stack-data==0.6.3",
    "tenacity==9.1.2",
    "thrift==0.22.0",
    "tiktoken==0.12.0",
    "tornado==6.5.4",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uuid-utils==0.13.0",
    "wcwidth==0.2.14",
    "wrapt==1.17.3",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
